requirements.txt
Flask
pillow
keras
gevent
requests
tensorflow
numpy
matplotlib
opencv-python